[WIP] Deprecate reexporter
This deprecates reexporter so that we can introduce a "better" addon structure.  Currently we autogenerate a "reexport" file that can not be statically looked up when resolving a depenency graph. The migration would be to have developers to include their own "reexport" file in the addon if it's necessary. The `DeprecatedReexporter` warns developers and then points them to a MIGRATION.md. The migration file does not exist yet.